% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{cats2h5}
\alias{cats2h5}
\title{A cats2h5 fonction for convert csv file to h5 file}
\usage{
cats2h5(filecsv = "", accres = 50, fileh5 = "")
}
\arguments{
\item{filecsv}{A input cats csv file.}

\item{accres}{input resolution}

\item{fileh5}{A output h5 data file.}
}
\description{
A cats2h5 fonction for convert csv file to h5 file
}
