% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{catsn2h5}
\alias{catsn2h5}
\title{A catsn2h5 fonction for convert csv file to h5 file}
\usage{
catsn2h5(filecsv = "", filecsvres = 50, fileh5 = "")
}
\arguments{
\item{filecsv}{A input cats csv file.}

\item{filecsvres}{input resolution}

\item{fileh5}{A output h5 data file.}
}
\description{
A catsn2h5 fonction for convert csv file to h5 file
}
