% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{demo_gui}
\alias{demo_gui}
\title{A demow_gui function for lunch a R-shiny application to plot datalogger view}
\usage{
demo_gui()
}
\description{
A demow_gui function for lunch a R-shiny application to plot datalogger view
}
