% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{democats2h5}
\alias{democats2h5}
\title{A democats2h5 fonction build demo cats h5 file}
\usage{
democats2h5(fileh5 = "", nbrow = 10000)
}
\arguments{
\item{fileh5}{imput data h5 file}

\item{nbrow}{number of row}
}
\description{
A democats2h5 fonction build demo cats h5 file
}
