% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{democatsmkbe}
\alias{democatsmkbe}
\title{A democatsmkbe fonction for generate ramdom data}
\usage{
democatsmkbe(fbe = "", nbrow = 10, nbseq = 2)
}
\arguments{
\item{fbe}{A outout be csv file.}

\item{nbrow}{input number of data rate in 1 seconde}

\item{nbseq}{input sequence lenght}
}
\description{
A democatsmkbe fonction for generate ramdom data
}
