% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{demolul2h5}
\alias{demolul2h5}
\title{A demolul2h5 function build demo lul h5 file}
\usage{
demolul2h5(fileh5 = "", nbrow = 10000)
}
\arguments{
\item{fileh5}{A h5 data file.}

\item{nbrow}{number of row}
}
\description{
A demolul2h5 function build demo lul h5 file
}
