% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{demowacu2h5}
\alias{demowacu2h5}
\title{A demowacu2h5 function build demo cats h5 file}
\usage{
demowacu2h5(fileh5 = "", nbrow = 10000)
}
\arguments{
\item{fileh5}{A h5 data file.}

\item{nbrow}{number of row}
}
\description{
A demowacu2h5 function build demo cats h5 file
}
