% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{LoggerList-class}
\alias{LoggerList-class}
\alias{LoggerList}
\title{A LoggerList reference class}
\description{
A LoggerList reference class
}
\section{Methods}{

\describe{
\item{\code{add(node)}}{add new node in the list.}

\item{\code{draw()}}{draw the objec value
\subsection{Return Value}{returns a list of String object representing the value}}
}}

