% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{LoggerLul-class}
\alias{LoggerLul-class}
\alias{LoggerLul}
\title{A LoggerLul reference class}
\description{
A LoggerLul reference class
}
\section{Methods}{

\describe{
\item{\code{draw()}}{draw the objec value
\subsection{Return Value}{returns a String object representing the value}}

\item{\code{h5init()}}{verify if h5 is correct version}

\item{\code{initmetriclst()}}{set metric list for this logger class}
}}

