% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggerui.R
\docType{class}
\name{LoggerUI-class}
\alias{LoggerUI-class}
\alias{LoggerUI}
\title{A LoggerUI reference class}
\description{
A LoggerUI reference class
}
\section{Fields}{

\describe{
\item{\code{loglst}}{list of logger class}

\item{\code{id}}{id of curent loger view}

\item{\code{ldatestart}}{curent start date}

\item{\code{nbrow}}{courent row number}
}}

\section{Methods}{

\describe{
\item{\code{gui()}}{plot logger list}
}}

