% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{lul2h5}
\alias{lul2h5}
\title{A lul2h5 function for concert lul csv file to h5 file}
\usage{
lul2h5(filecsv = "", accres = 25, fileh5 = "")
}
\arguments{
\item{filecsv}{A input LUL csv file.}

\item{accres}{input number of data rate in 1 seconde}

\item{fileh5}{A output h5 data file.}
}
\description{
A lul2h5 function for concert lul csv file to h5 file
}
