% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{Metric-class}
\alias{Metric-class}
\alias{Metric}
\title{Metric reference class}
\description{
Metric reference class
}
\section{Fields}{

\describe{
\item{\code{name}}{title metric in chart}

\item{\code{colid}}{start column id}

\item{\code{connb}}{number of column for this metric}
}}

\section{Methods}{

\describe{
\item{\code{draw()}}{draw the objec value
\subsection{Return Value}{returns a String object representing the value}}
}}

