% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{MetricList-class}
\alias{MetricList-class}
\alias{MetricList}
\title{A MetricList reference class}
\description{
A MetricList reference class
}
\section{Methods}{

\describe{
\item{\code{draw()}}{draw the objec value
\subsection{Return Value}{returns a list of String object representing the value}}

\item{\code{getsize()}}{return lenght of element.
\subsection{Return Value}{returns a non-negativ numeric}}

\item{\code{slctset(v)}}{enable or disable metric view
\subsection{Parameters}{\itemize{
  \item{\code{v} True or False vector}
}}}
}}

