% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{wacu2h5}
\alias{wacu2h5}
\title{A wacu2h5 fonction for concert wacu csv file to h5 file}
\usage{
wacu2h5(filewacucsv = "", fileh5 = "")
}
\description{
A wacu2h5 fonction for concert wacu csv file to h5 file
}
