% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{wacu2hacc2}
\alias{wacu2hacc2}
\title{A wacu2hacc2 function for insert wacu acc csv file to h5 file}
\usage{
wacu2hacc2(filewacucsv = "", fileh5 = "", size = 11274058,
  accfreq = 25)
}
\description{
A wacu2hacc2 function for insert wacu acc csv file to h5 file
}
