% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{wacu2hacc3}
\alias{wacu2hacc3}
\title{A wacu2hacc3 function for insert wacu acc csv file to h5 file}
\usage{
wacu2hacc3(filewacucsv = "", fileh5 = "", size = 11274058,
  accfreq = 25)
}
\description{
A wacu2hacc3 function for insert wacu acc csv file to h5 file
}
