% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{wacupre2h5}
\alias{wacupre2h5}
\title{A wacupre2h5 function for concert a preview from wacu csv file to h5 file}
\usage{
wacupre2h5(filecsv = "", fileacccsv = "", fileh5 = "")
}
\arguments{
\item{filecsv}{A input WACU csv file.}

\item{fileacccsv}{A input WACU ACC csv file.}

\item{fileh5}{A output h5 data file.}
}
\description{
A wacupre2h5 function for concert a preview from wacu csv file to h5 file
}
