% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggerui.R
\docType{class}
\name{ZoomHistory-class}
\alias{ZoomHistory-class}
\alias{ZoomHistory}
\title{A ZoomHistory reference class}
\description{
A ZoomHistory reference class
}
\section{Methods}{

\describe{
\item{\code{draw()}}{draw the objec value
\subsection{Return Value}{returns a matrix of value}}

\item{\code{pop()}}{pop one history position}

\item{\code{push(s, e)}}{push new history position in array.}
}}

