% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{demoaxytrek2h5}
\alias{demoaxytrek2h5}
\title{A demoaxytrek2h5 fonction build demo cats h5 file}
\usage{
demoaxytrek2h5(fileh5 = "", nbrow = 10000)
}
\arguments{
\item{fileh5}{input data H5 file}

\item{nbrow}{number of row}
}
\description{
A demoaxytrek2h5 fonction build demo cats h5 file
}
\examples{
demoaxytrek2h5("demoaxy.h5",nbrow=100)
}
