% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{Logger-class}
\alias{Logger-class}
\alias{Logger}
\title{A Logger reference class}
\description{
A Logger reference class
}
\section{Fields}{

\describe{
\item{\code{filedata}}{nom du fichier de donnée}

\item{\code{filebehavior}}{nom du chier des comportement}
}}


\author{
sebastien geiger
}
