% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{lul2h5}
\alias{lul2h5}
\title{A lul2h5 function for concert lul csv file to h5 file}
\usage{
lul2h5(filecsv = "", fileh5 = "", sep = "\\t")
}
\arguments{
\item{filecsv}{A input LUL csv file.}

\item{fileh5}{A output h5 data file.}

\item{sep}{input the field separator character.}
}
\description{
A lul2h5 function for concert lul csv file to h5 file
}
