% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{wacu2h5}
\alias{wacu2h5}
\title{A wacu2h5 function for concert wacu csv file to h5 file}
\usage{
wacu2h5(filecsv = "", fileh5 = "", rtctick = 1, accres = 50,
  datestartstring = "")
}
\arguments{
\item{filecsv}{A input WACU csv file.}

\item{fileh5}{A output h5 data file.}

\item{rtctick}{tpl frequence}

\item{accres}{acc frequence}

\item{datestartstring}{A Date string in GMT}
}
\description{
A wacu2h5 function for concert wacu csv file to h5 file
}
