% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{wacu2h5old}
\alias{wacu2h5old}
\title{A wacu2h5old fonction for concert wacu csv file to h5 file}
\usage{
wacu2h5old(filewacucsv = "", fileh5 = "")
}
\arguments{
\item{filewacucsv}{A input WACU csv file.}

\item{fileh5}{A output h5 data file.}
}
\description{
A wacu2h5old fonction for concert wacu csv file to h5 file
}
